requirements.txt
certifi==2020.12.5
kaggle==1.5.8
numpy==1.19.1
Pillow==7.2.0
python-dateutil==2.8.1
python-slugify==4.0.1
slugify==0.0.1
text-unidecode==1.3
urllib3==1.24.3
